Add validations to CountryInfoDtoValidator, rename folders and files
