+++ src/Services/Identity/Identity.Application/DTO/CountryInfo/CountryInfoDtoValidator.cs
@@ -1,8 +1,23 @@
 using FluentValidation;
+using Identity.Domain.AggregationModels.ApplicationUser;
+using Identity.Domain.AggregationModels.ApplicationUser.ValueObjects;
 
 namespace Identity.Application.DTO;
 
 public class CountryInfoDtoValidator: AbstractValidator<CountryInfoDto>
 {
-    
+    public CountryInfoDtoValidator()
+    {
+        RuleFor(c => c.Name)
+            .NotNull()
+            .WithMessage("Country cannot be null.");
+
+        RuleFor(c => c.ISO)
+            .NotNull()
+            .WithMessage("Country ISO cannot be empty.")
+            .Length(CountryInfoEntityValidationConstants.CountryISOCodeLength)
+            .WithMessage("Country ISO has to be in three characters length.")
+            .Matches("^[A-Z]{3}$")
+            .WithMessage("Country ISO is invalid.");
+    }
 }
