+++ src/Services/Identity/Identity.Application/Mappers/UserMapper/ApplicationUserMapper.cs
@@ -1,5 +1,5 @@
 using Identity.Application.DTO;
-using Identity.Application.DTO.Auth;
+using Identity.Application.DTO.RegisteringUser;
 using Identity.Domain.AggregationModels.ApplicationUser;
 using Identity.Domain.AggregationModels.ApplicationUser.ValueObjects;
